refactor: Move GitRevision.BuildSummary() out
