+++ UnitTests/GitCommandsTests/GitRevisionTests.cs
@@ -26,31 +26,5 @@ namespace GitCommandsTests
             Assert.False(GitRevision.IsFullSha1Hash("0000000000000000000000000000000000000000 "));
             Assert.False(GitRevision.IsFullSha1Hash(" 0000000000000000000000000000000000000000"));
         }
-
-        [TestCase(null)]
-        [TestCase("toto")]
-        [TestCase("toto\ntata\ntiti")]
-        public void Should_have_same_content_When_no_ellipsis(string bodyContent)
-        {
-            var gitRevision = new GitRevision(ObjectId.Random());
-            gitRevision.Body = bodyContent;
-            gitRevision.HasMultiLineMessage = true;
-            Assert.AreEqual(bodyContent, gitRevision.BodySummary);
-        }
-
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        [Test]
-        [TestCase("Too_many_lines")]
-        [TestCase("Too_long_lines")]
-        public void Should_do_ellipsis(string testName)
-        {
-            var content = EmbeddedResourceLoader.Load(Assembly.GetExecutingAssembly(), $"{GetType().Namespace}.MockData.{testName}.txt");
-            var gitRevision = new GitRevision(ObjectId.Random()) { Body = content, HasMultiLineMessage = true };
-
-            using (ApprovalResults.ForScenario(testName))
-            {
-                Approvals.Verify(gitRevision.BodySummary);
-            }
-        }
     }
 }
